requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
urllib3==2.0.6
python-dotenv>=1.0.0
discord.py>=2.3.1
beautifulsoup4>=4.12.2
bs4>=0.0.1